Address Greptile: don't fail a landed write, and keep fcntl off the config path
Two findings on #345, both real.

**A post-rename fsync failure reported a write that landed as failed (P1).** The
parent-directory fsync runs after `replace`, so propagating its error told the
caller nothing was written when the file already held the new contents. A caller
acting on that (reporting a failed promotion, retrying, rolling back) would act
on the opposite of the truth. Worse than misleading: some FUSE, SMB and NFS
mounts reject `fsync` on a directory fd unconditionally, which turned every
write on such a mount into an error for a file written correctly every time.

The previous commit documented this in `Raises:` instead of fixing it, which was
the wrong call. `_fsync_directory` now logs the failure and returns, so an
exception from `write_bytes_atomic` again means exactly one thing: the write did
not land. What is genuinely lost is the proof of durability against a power
loss, not the write, and that is what the warning says. The test that asserted
the old behaviour now asserts the new one, plus one that the warning is emitted:
not raising must not mean saying nothing.

**`fcntl` reached the import path of almost every command (P2).** Putting the
lock and the atomic writers in one module meant `wmo/config/config.py` imported
`fcntl`, and on main nothing central did: the only unconditional `import fcntl`
was in `pool.py`, which `wmo/providers/__init__.py` does not import. Everything
that only writes a file whole (`.wmo/config.toml`, `.wmo/settings.toml`, a model
card, a fitted policy, an outcome matrix, an endpoint dial) needs no locking at
all.

Split into `wmo/core/locks.py`, which is the half that needs `fcntl`. A lazy
import would have been the smaller diff but AGENTS.md rule 8 forbids one for
optional convenience, and the split is the honest fix: these are two contracts,
not one with a flag. Verified by import: bare `import wmo`, `wmo.config` and
`wmo/config/card.py` no longer load `wmo.core.locks`, while `pool.py` and both
stores still do. `wmo.serving` and `wmo.optimize.outcomes` still reach it, via
their own pre-existing imports of `pool`.

Note this does not make wmo run on Windows and is not claimed to; it stops the
atomic-write half from being the reason it cannot.
